Added file and line-number information to reports.
This is a pretty massive commit. The plumbing all the way from assertion up was converted to use Report objects. Those Report objects were extended to hold all the data, and the logic of how to format them was pushed all the way into the Reporter.

The trickiest thing here is how the location objects get created. Go's runtime.Caller() takes a depth, which is how many levels up the stack should be traversed. Since the binfile wraps It() with another layer of indirection, we have to adjust the number of frames we go up in order to find the right one for pending specs...but only from the binfile.