Make a distinction between errors and failures.
In a library like rspec, exceptions happen and cause an error. That's different than an assertion failure. In Go, we don't have exceptions, but we do have the comma-err/ok patterns to indicate the same sort of thing. Context.Error is there to report those, and so it should produce a different kind of report to distinguish them from assertion failures.