DRYed up reporters.
The testingReporter was duplicating code that was also in dotReporter. I essentially moved it into the codebase as basicReporter and then also moved the interfaces into the codebase. This means that not only did I DRY up my code, but I'm also actually testing more of the code that gets run for real. The only behavior left in dotReporter is actually printing stuff.