Added specdoc runner like rspec's.
The command-line parsing happens in bin/specify and simply changes the reporter used in the generated _specify_.go.

Since there are now two ReportSummary implementations that produce output, I refactored things so that there's an output strategy for the things that are different and an outputReporter that uses it but does not duplicate the summary stuff at the end.

Thanks to Andrea Fazzi <andrea.fazzi@alcacoop.it> for the original contribution.